useAuth/Token и добавил preloader
