+++ src/components/Preloader/Preloader.jsx
@@ -1,0 +1,14 @@
+import { useEffect, useState } from 'react'
+import './Preloader.css'
+
+const Preloader = () => {
+  return (
+    <div className="preloader">
+      <div className="preloader__container">
+        <span className="preloader__round"></span>
+      </div>
+    </div>
+  )
+};
+
+export default Preloader;
